- metro/metro.cpp: fixed regression for puzzlet
- technos/battlane.cpp, technos/blockout.cpp, technos/wwfsstar.cpp: consolidated drivers in single files, very minor cleanups
